requirements.txt
fastapi
uvicorn
sqlalchemy
deepl
python-dotenv